[添加短信登录页面]

+++ lib/themes/theme_size.dart
@@ -1,3 +1,3 @@
-int defaultActionBarHeight = 56;
-int defaultStatusBarHeight = 20;
+double defaultActionBarHeight = 56;
+double defaultStatusBarHeight = 20;
 double defaultBottomNavigationBarHeight = 56;
